requirements.txt
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
watchdog>=4.0.0
flask>=3.0.0
